feat: add exclude option
This option can be used multiple times to exclude keys from the
extraction / report for those situations where the 'suspected dynamic
keys' logic is not enough and the user prefers to drop known keys.

It only makes sense for the unused keys, therefore it does not affect
the missing keys.

Link: https://github.com/pixari/vue-i18n-extract/issues/141