refactor: only package necessary files
Adds the `files` property to the package.json file including the paths `bin` and `dist` which contain the essential distribution files of the package. Previously, the entire source code was packaged with the files uploaded to npm (for comparison, run `npm pack` with and without this change). This unnecessarily bloats a consuming project’s dependency storage (e.g. `node_modules`).