feat: add outputOrder option
Adds a new optional config and CLI option `outputOrder` which when set to `'lexical'` sorts reports and language files (when using the `add` option) alphabetically. By default, the option is set to `'append'` which represents the current behavior of appending new translation keys to a language file when using the `add` option.